--- firebase-test/functions.js
@@ -1,3 +1,0 @@
-export function add(a, b) {
-  return a + b;
-}
