--- firebase-test/index.css
@@ -1,45 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-body {
-  background-color: aliceblue;
-}
-
-.container {
-  border: solid 2px black;
-  display: flex;
-  flex-direction: column;
-  max-width: 320px;
-  margin: 30px auto;
-  padding: 20px;
-}
-
-img {
-  width: 150px;
-  margin: 0 auto;
-  object-fit: contain;
-}
-
-input {
-  color: #432000;
-  background-color: #dce1eb;
-  border: 0;
-  padding: 15px;
-  text-align: center;
-  font-size: 20px;
-  border-radius: 8px;
-  margin: 10px 0;
-}
-button {
-  color: #dedede;
-  background-color: #c3627c;
-  border: 0;
-  padding: 15px;
-  text-align: center;
-  font-size: 20px;
-  border-radius: 8px;
-  cursor: pointer;
-}
