--- firebase-test/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="index.css">
-</head>
-
-<body>
-    <div class="container">
-        <img src="https://images.unsplash.com/photo-1528018794703-b80141aad05b?w=400&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MjB8fGNoaXBtdW5rfGVufDB8fDB8fHww"
-            alt="">
-        <input type="text" id="input-field" placeholder="Bread">
-        <button id="add-button">go</button>
-    </div>
-    <script src="apikey.js"></script>
-    <script type="module" src="functions.js"></script>
-    <script type="module" src="index.js"></script>
-</body>
-
-</html>
