+++ question.html
@@ -28,7 +28,9 @@
       <button id="answer-right" class="answer">무슨 맛 인데</button>
     </div>
   </div>
-  <script src="/scripts/question.js"></script>
+
+  <script src="/scripts/apikey.js"></script>
+  <script type="module" src="/scripts/question.js"></script>
 </body>
 
 </html>
