+++ scripts/question.js
@@ -1,10 +1,47 @@
+import { initializeApp } from "https://www.gstatic.com/firebasejs/9.15.0/firebase-app.js";
+import {
+  getFirestore,
+  collection,
+  getDocs,
+  doc,
+  updateDoc,
+} from "https://www.gstatic.com/firebasejs/9.15.0/firebase-firestore.js";
+
+const API_KEY = config.apikey;
+const firebaseConfig = {
+  apiKey: API_KEY,
+  authDomain: "oz-project-16f76.firebaseapp.com",
+  projectId: "oz-project-16f76",
+  storageBucket: "oz-project-16f76.appspot.com",
+  messagingSenderId: "603853071891",
+  appId: "1:603853071891:web:8f5897ccba3d96e7e541e6",
+};
+
+const appFS = initializeApp(firebaseConfig);
+const db = getFirestore(appFS);
+const usersRef = collection(db, "users");
+let bufCount;
+
+async function getCount() {
+  const querySnapshot = await getDocs(collection(db, "users"));
+  querySnapshot.forEach((doc) => {
+    bufCount = Number(doc.data().count);
+    bufCount++;
+    console.log(bufCount);
+    updateCount();
+  });
+}
+
+async function updateCount() {
+  const userRefUpdate = doc(db, "users", "0");
+
+  // Set the "capital" field of the city 'DC'
+  await updateDoc(userRefUpdate, {
+    count: bufCount,
+  });
+}
+
 const question = document.getElementById("question");
-const answer = document.querySelector(".answers");
-const bread1 = document.querySelector(".bread1");
-const bread2 = document.querySelector(".bread2");
-const bread3 = document.querySelector(".bread3");
-const bread4 = document.querySelector(".bread4");
-const bread5 = document.querySelector(".bread5");
 const answerTop = document.querySelector("#answer-left");
 const answerBottom = document.querySelector("#answer-right");
 
@@ -124,29 +161,6 @@ const questions = [
   },
 ];
 
-const breadResult = [
-  {
-    breadType: "팥붕어빵",
-    comment: "이것은 팥붕어빵 입니다. ",
-  },
-  {
-    breadType: "슈붕어빵",
-    comment: "이것은 슈붕어빵 입니다. ",
-  },
-  {
-    breadType: "피자붕어빵",
-    comment: "이것은 피자붕어빵 입니다. ",
-  },
-  {
-    breadType: "녹두붕어빵",
-    comment: "이것은 녹두붕어빵 입니다. ",
-  },
-  {
-    breadType: "고구마붕어빵",
-    comment: "이것은 고구마붕어빵 입니다. ",
-  },
-];
-
 //답변 클릭했을 때 배열을 하나씩 순회하면서 질문이랑 답 텍스트가 바뀌게 + 콘솔에 붕어빵 종류 담기게
 let ArrIndex = 1;
 const typeArr = []; //type 담아주는 배열 -> 최종적으로 15개 담깁니다!
@@ -160,10 +174,8 @@ answerTop.addEventListener("click", function () {
   } else {
     const answerAType = questions[ArrIndex - 1].answers.A.type;
     typeArr.push(answerAType); // 배열에 어떤 붕어빵인지 넣어주기
-
     location.href = "result.html";
-    let arrMaxIndex = maxBreadIndex(typeArr);
-    printResult(arrMaxIndex);
+    maxBreadIndex(typeArr);
   }
 });
 
@@ -177,8 +189,7 @@ answerBottom.addEventListener("click", function () {
     const answerBType = questions[ArrIndex - 1].answers.B.type;
     typeArr.push(answerBType); // 배열에 어떤 붕어빵인지 넣어주기
     location.href = "result.html";
-    let arrMaxIndex = maxBreadIndex(typeArr);
-    printResult(arrMaxIndex);
+    maxBreadIndex(typeArr);
   }
 });
 
@@ -207,6 +218,8 @@ function maxBreadIndex(arr) {
   let maxIndex = answerArr.indexOf(Math.max(...answerArr));
   console.log(answerArr);
   console.log(maxIndex);
-  return localStorage.setItem("breadType", maxIndex);
 
+  return localStorage.setItem("breadType", maxIndex);
 }
+
+getCount();
